requirements.txt
asgiref==3.4.1
async-timeout==4.0.2
authentication==1.1.0
boto3==1.23.10
botocore==1.26.10
cffi==1.15.1
click==8.0.4
cryptography==37.0.4
dataclasses
Deprecated==1.2.13
Django==3.2.14
django-phonenumber-field==6.0.0
djangorestframework==3.13.1
Flask==2.0.3
importlib-metadata==4.8.3
itsdangerous==2.0.1
Jinja2==3.0.3
jmespath==0.10.0
jwt==1.3.1
MarkupSafe==2.0.1
packaging==21.3
passlib==1.7.4
phonenumbers==8.12.51
phonenumberslite==8.12.51
Pillow==8.4.0
pycparser==2.21
pycryptodome==3.15.0
PyJWT==2.4.0
PyMySQL==1.0.2
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
redis==4.3.4
s3transfer==0.5.2
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.1.1
urllib3==1.26.10
Werkzeug==2.0.3
wrapt==1.14.1
WTForms==3.0.0
zipp==3.6.0
